% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_idml.R
\name{check_idml}
\alias{check_idml}
\alias{check_idml_resource}
\title{Check IDML objects}
\usage{
check_idml_resource(
  idml,
  resource = NULL,
  arg = caller_arg(idml),
  call = caller_env()
)
}
\arguments{
\item{idml}{A \code{idml} class object created with \code{\link[=read_idml]{read_idml()}}.}

\item{resource}{One or more resource names: "Graphic.xml", "Fonts.xml",
or "Styles.xml"}
}
\description{
\code{\link[=check_idml_resource]{check_idml_resource()}} helps check if an \code{idml} object contains the expected
resource file.
}
\keyword{internal}
