% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_read_settings.R
\name{id_read_settings}
\alias{id_read_settings}
\title{Read exported Adobe InDesign user application settings}
\usage{
id_read_settings(path, format = "data.frame")
}
\arguments{
\item{path}{Path to existing file with exported user application settings.}

\item{format}{Output format. Must be "data.frame"}
}
\description{
Adobe InDesign user application settings can be exported to a SQLite database
containing XML files. This function reads the
}
\keyword{internal}
