% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idml_contents.R
\name{idml_contents}
\alias{idml_contents}
\alias{get_idml_contents}
\title{Get contents from an IDML object}
\usage{
get_idml_contents(
  idml,
  dir = NULL,
  file = NULL,
  format = "xml_document",
  ns = character(),
  error_call = caller_env()
)
}
\arguments{
\item{idml}{A `idml` class object created with [read_idml()].}

\item{dir, file}{Directory and file name from IDML contents to return.}

\item{format}{"list" or "xml_document". If "list", contents are converted
with [xml2::as_list()] using the supplied `ns` parameter.}

\item{ns}{Optionally, a named vector giving prefix-url pairs, as produced
by \code{\link[xml2:xml_ns]{xml_ns()}}. If provided, all names will be explicitly
qualified with the ns prefix, i.e. if the element \code{bar} is defined
in namespace \code{foo}, it will be called \code{foo:bar}. (And
similarly for attributes). Default namespaces must be given an explicit
name. The ns is ignored when using \code{\link[xml2:xml_name<-]{xml_name<-()}} and
\code{\link[xml2:xml_set_name]{xml_set_name()}}.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
[get_idml_contents()] is a helper for extracting or converting `xml_document`
objects from the `idml` list object.
}
