% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idml_stories.R
\name{idml_list_stories}
\alias{idml_list_stories}
\alias{idml_list_story_paras}
\title{List stories from the content of an idml object}
\usage{
idml_list_stories(idml, format = "contents")

idml_list_story_paras(idml, ...)
}
\arguments{
\item{idml}{An \code{idml} class object.}

\item{format}{Format of output to return. Must be "contents"}
}
\value{
A bare named list with four elements: Stories, StoryPreference,
InCopyExportOption, and ParagraphStyleRange.
}
\description{
\code{\link[=idml_list_stories]{idml_list_stories()}} lists stories and related elements from a idml object.
\code{\link[=idml_list_story_paras]{idml_list_story_paras()}} returns just the ParagraphStyleRange element of the
stories.
}
\examples{
\dontrun{
if(interactive()){
  path <- system.file("idml/letter_portrait_standard.idml", package = "idmlr")

  idml_obj <- read_idml(path)

  idml_list_stories(idml_obj)
 }
}
}
