% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idml_resource.R
\name{idml_list_styles}
\alias{idml_list_styles}
\title{List styles from an idml object}
\usage{
idml_list_styles(..., type = "paragraph", format = "list")
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_idml_styles]{get_idml_styles}}
  \describe{
    \item{\code{idml}}{A \code{idml} class object created with \code{\link[=read_idml]{read_idml()}}.}
    \item{\code{parent_nm}}{Name of top-level parent node. If parent_nm is a
character vector, \code{\link[=extract_doc_df]{extract_doc_df()}} returns a data frame list.}
  }}

\item{type}{Type of data to extract from chlildren nodes: "attr" (attributes)
or "text".}

\item{format}{Output format must be "list"}
}
\value{
A bare list with four elements: a named list of style attributes
(names are style names), properties for the style XML nodes, attributes for
the style XML nodes, and the contents of each XML node.
}
\description{
\code{\link[=idml_list_styles]{idml_list_styles()}} takes an idml object and returns a bare list with names,
properties, attributes, and content from the XML nodes definition the
paragraph, character, or other styles of the specified type.
}
\examples{
\dontrun{
if(interactive()){
  path <- system.file("idml/letter_portrait_standard.idml", package = "idmlr")

  idml_obj <- read_idml(path)

  idml_list_styles(idml_obj)
 }
}
}
