% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idml_resource.R
\name{idml_resource}
\alias{idml_resource}
\alias{get_idml_resource}
\alias{get_idml_fonts}
\alias{get_idml_styles}
\alias{get_idml_graphic}
\alias{get_idml_preferences}
\title{Get fonts, styles, and graphics from an IDML object}
\usage{
get_idml_resource(
  idml,
  resource,
  format = "xml_document",
  ...,
  error_call = caller_env()
)

get_idml_fonts(idml, format = "xml_document", ...)

get_idml_styles(idml, format = "xml_document", ...)

get_idml_graphic(idml, format = "xml_document", ...)

get_idml_preferences(idml, format = "xml_document", ...)
}
\arguments{
\item{idml}{A `idml` class object created with [read_idml()].}

\item{resource}{One or more resource names: "Graphic.xml", "Fonts.xml",
or "Styles.xml"}

\item{format}{"list" or "xml_document". If "list", contents are converted
with [xml2::as_list()] using the supplied `ns` parameter.}

\item{...}{
  Arguments passed on to \code{\link[=get_idml_contents]{get_idml_contents}}
  \describe{
    \item{\code{ns}}{Optionally, a named vector giving prefix-url pairs, as produced
by \code{\link[xml2:xml_ns]{xml_ns()}}. If provided, all names will be explicitly
qualified with the ns prefix, i.e. if the element \code{bar} is defined
in namespace \code{foo}, it will be called \code{foo:bar}. (And
similarly for attributes). Default namespaces must be given an explicit
name. The ns is ignored when using \code{\link[xml2:xml_name<-]{xml_name<-()}} and
\code{\link[xml2:xml_set_name]{xml_set_name()}}.}
  }}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Get fonts, styles, and graphics from an IDML object

[get_idml_resource()] extracts a resource from an `idml` object in the
supplied format. [get_idml_fonts()], [get_idml_styles()],
[get_idml_graphic()], and [get_idml_preferences()] are helpers for extracting
the Fonts.xml, Styles.xml, Graphic.xml, and Preferences.xml files.
}
