% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idml_spreads.R
\name{idml_spreads}
\alias{idml_spreads}
\alias{get_idml_spreads}
\title{Get spreads from an IDML object}
\usage{
get_idml_spreads(
  idml,
  format = "list",
  names_to = "Spread",
  ...,
  .error_call = caller_env()
)
}
\arguments{
\item{idml}{A `idml` class object created with [read_idml()].}

\item{format}{"list", "xml_document", or "data.frame". If "list", contents
are converted with [xml2::as_list()] using the supplied `ns` parameter.}

\item{names_to}{By default, \code{names(x)} are lost. To keep them, supply a
string to \code{names_to} and the names will be saved into a column with that
name. If \code{names_to} is supplied and \code{x} is not named, the position of
the elements will be used instead of the names.}

\item{...}{
  Arguments passed on to \code{\link[=get_idml_contents]{get_idml_contents}}
  \describe{
    \item{\code{ns}}{Optionally, a named vector giving prefix-url pairs, as produced
by \code{\link[xml2:xml_ns]{xml_ns()}}. If provided, all names will be explicitly
qualified with the ns prefix, i.e. if the element \code{bar} is defined
in namespace \code{foo}, it will be called \code{foo:bar}. (And
similarly for attributes). Default namespaces must be given an explicit
name. The ns is ignored when using \code{\link[xml2:xml_name<-]{xml_name<-()}} and
\code{\link[xml2:xml_set_name]{xml_set_name()}}.}
    \item{\code{parent_nm}}{Name of top-level parent node. If parent_nm is a
character vector, [extract_doc_df()] returns a data frame list.}
    \item{\code{unique_nm}}{If `TRUE`, ensure that parent_node names are unique before
using them to extract data from the `xml_document`. Ignored if `node_names`
is supplied.}
    \item{\code{type}}{Type of data to extract from chlildren nodes: "attr" (attributes)
or "text".}
    \item{\code{error_call}}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
  }}

\item{.error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Get spreads from an IDML object

[get_idml_spreads()] extracts a list of xml documents for spreads or a data
frame from a `idml` object.
}
