% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idml_stories.R
\name{idml_stories}
\alias{idml_stories}
\title{Pull the stories element from an idml object}
\usage{
idml_stories(idml, error_call = caller_env())
}
\arguments{
\item{idml}{An \code{idml} class object.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=idml_stories]{idml_stories()}} pulls the "Stories" elements from the "contents" of an idml
object.
}
\keyword{internal}
