% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is_list_of_xml_documents}
\alias{is_list_of_xml_documents}
\title{Is the object a list of XML document objects?}
\usage{
is_list_of_xml_documents(x)
}
\description{
Is the object a list of XML document objects?
}
