% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-extract.R
\name{xml_doc_to_df}
\alias{xml_doc_to_df}
\title{Create a data frame from an XML document}
\usage{
xml_doc_to_df(
  doc = NULL,
  parent_nm = NULL,
  parent_node = NULL,
  node_names = NULL,
  unique_nm = TRUE,
  nested = TRUE,
  type = "attr",
  ...
)
}
\arguments{
\item{doc}{A XML document.}

\item{parent_nm}{Name of top-level parent node. If parent_nm is a
character vector, [extract_doc_df()] returns a data frame list.}

\item{parent_node}{A XML document, node, or node set.}

\item{node_names}{Node names to iterate over. Default `NULL`. Typically
created from children of parent node.}

\item{unique_nm}{If `TRUE`, ensure that parent_node names are unique before
using them to extract data from the `xml_document`. Ignored if `node_names`
is supplied.}

\item{nested}{If `TRUE`, use `parent_nm` to extract a parent node from doc.
If `FALSE`, `parent_node` defaults to doc (unless `parent_node` is supplied).}

\item{type}{Type of data to extract from chlildren nodes: "attr" (attributes)
or "text".}

\item{...}{
  Arguments passed on to \code{\link[readr:type_convert]{readr::type_convert}}
  \describe{
    \item{\code{df}}{A data frame.}
    \item{\code{col_types}}{One of \code{NULL}, a \code{\link[readr:cols]{cols()}} specification, or
a string. See \code{vignette("readr")} for more details.

If \code{NULL}, column types will be imputed using all rows.}
    \item{\code{na}}{Character vector of strings to interpret as missing values. Set this
option to \code{character()} to indicate no missing values.}
    \item{\code{trim_ws}}{Should leading and trailing whitespace (ASCII spaces and tabs) be trimmed from
each field before parsing it?}
    \item{\code{locale}}{The locale controls defaults that vary from place to place.
The default locale is US-centric (like R), but you can use
\code{\link[readr:locale]{locale()}} to create your own locale that controls things like
the default time zone, encoding, decimal mark, big mark, and day/month
names.}
    \item{\code{guess_integer}}{If \code{TRUE}, guess integer types for whole numbers, if
\code{FALSE} guess numeric type for all numbers.}
  }}
}
\description{
Based on https://rud.is/rpubs/xml2power/
}
\keyword{internal}
