% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_idml.R
\name{read_idml}
\alias{read_idml}
\title{Read an InDesign Markup Language (IDML) file to a `idml` object}
\usage{
read_idml(file, ..., exdir = tempdir())
}
\arguments{
\item{file}{Path to a file ending in a `.idml` file extension.}

\item{...}{Additional parameters passed to [xml2::read_xml()]}

\item{exdir}{Directory to extract files to when using [base::unzip()],
Default: `tempdir()`.}
}
\value{
A `idml` class object.
}
\description{
[read_idml()] takes a file path for an `idml` file and returns an `idml`
object: a list containing a file name, file path, and the contents of the
file as a list of `xml_document` objects and character vectors (for the
MIMETYPE file only).
}
\seealso{
[xml2::read_xml()]
}
